// src/app/models/ConfigEquiposAlertas.ts
export class ConfigEquiposAlertas {
    email: string;
    Rcada: number;
    tiempo: number;
    email2: string;
    Rcada2: number;
    tiempo2: number;
    nombre1: number;
    nombre2: number;
    nombre3: number;
    nombre4: number;
    nombre5: number;
    avisarCorreo: boolean;
    activaN1: boolean;
    activaN2: boolean;
    activaN3: boolean;
    activaN4: boolean;
    activaN5: boolean;
    _id: string;
}